Moved things into Model/controller designations, fixed registration activity bug by registering all activities
