+++ app/src/main/AndroidManifest.xml
@@ -11,15 +11,19 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".controllers.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".login" />
-        <activity android:name=".Account"></activity>
+        <activity android:name=".controllers.login" />
+        <activity android:name=".controllers.Account"/>
+        <activity android:name=".controllers.registration"/>
+        <activity android:name=".controllers.RegisteredActivity"/>
+
+
     </application>
