+++ app/src/main/java/com/example/sotw/donationtracker/controllers/Account.java
@@ -1,10 +1,12 @@
-package com.example.sotw.donationtracker;
+package com.example.sotw.donationtracker.controllers;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.sotw.donationtracker.R;
 import com.google.firebase.auth.FirebaseAuth;
 
 public class Account extends AppCompatActivity {
