+++ app/src/main/res/layout/activity_registered.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegisteredActivity">
+    tools:context=".controllers.RegisteredActivity">
 
     <TextView
         android:id="@+id/textView"
