+++ app/src/main/res/layout/activity_registration.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".registration">
+    tools:context=".controllers.registration">
 
     <TextView
         android:id="@+id/textView6"
